Added Testing (#21)
* Added Testing
Same as from main bot, debug entry should also be identical

* Make tests run
Before they weren't running, not too sure why, I moved them into a folder to create a package for the test, and it works now.

In main bot it was like how the current codebase was in the previous commit. Unsure why this was needed, but at least it works now.